refactor: types.tsをドメイン別に分割
- types/photo.ts: 写真・解析結果関連の型
- types/analysis.ts: 解析問題・お手本・履歴関連
- types/learning.ts: 学習ルール・GitHub同期関連
- types/storage.ts: ストレージ・設定関連
- types/usage.ts: API使用量・コスト関連
- types/index.ts: 後方互換性のため全型を再エクスポート

usageTrackerの型定義をtypes/usage.tsに移動し、
後方互換性を維持しながらモジュールを整理。
