+++ types/index.ts
@@ -1,0 +1,56 @@
+/**
+ * 型定義のメインエントリポイント
+ * 後方互換性のためすべての型を再エクスポート
+ */
+
+// Photo types
+export type {
+  PhotoMetadata,
+  AppMode,
+  PhotoCategory,
+  FieldChange,
+  ChangeStage,
+  AIAnalysisResult,
+  PhotoRecord,
+  ProcessingStats,
+  LogEntry,
+  SortPolicy,
+} from './photo';
+
+export { SORT_POLICIES } from './photo';
+
+// Analysis types
+export type {
+  AnalysisIssue,
+  IssueType,
+  IssueStatus,
+  AnalysisExample,
+  AnalysisSession,
+  AnalysisHistoryEntry,
+} from './analysis';
+
+// Learning types
+export type {
+  LearnedRule,
+  LearnedAlias,
+  LearnedExample,
+  LearnedSettings,
+} from './learning';
+
+// Storage types
+export type {
+  ProjectConfig,
+  PersonalConfig,
+  EncryptedCredentials,
+  SessionData,
+  SyncResult,
+  ConflictInfo,
+  StorageState,
+} from './storage';
+
+// Usage types
+export type {
+  UsageRecord,
+  UsageSummary,
+  CostEstimate,
+} from './usage';
